+++ source/geh_common/src/geh_common/domain/types/__init__.py
@@ -21,11 +21,9 @@ __all__ = [
     "QuantityQuality",
     "QuantityUnit",
     "SettlementMethod",
-    "QuantityQuality",
     "MeteringPointResolution",
     "ChargeQuality",
     "ChargeType",
     "ChargeUnit",
     "ChargeResolution",
-    "NetSettlementGroup",
 ]